1. 新增删除自己kids业务，目前该业务因为服务端接口未正式完成，暂时不可见 2. 删除kids后，使用数据库中现有数据更新缓存的focusKids

+++ Swing/data/src/main/java/com/kidsdynamic/data/net/kids/KidsApi.java
@@ -41,7 +41,7 @@ public interface KidsApi {
 //    400	Bad request. Missing some parameters
 //    500	Internal error. Please send me the error. I will fix i
     @DELETE("v1/kids/delete")
-    Call<Object> kidsDelete(@Query("kidId") int kidId);
+    Call<Object> kidsDelete(@Query("kidId") long kidId);
 
     //Retrieve kids belong to the signed in user
     @GET("v1/kids/list")
